OE-258 create test to validate if the level is correct

+++ database/migrations/2021_05_06_182808_create_user_enium_point_levels_table.php
@@ -16,7 +16,7 @@ class CreateUserEniumPointLevelsTable extends Migration
         Schema::create('user_enium_point_levels', function (Blueprint $table) {
             $table->id();
             $table->integer('level');
-            $table->decimal('point');
+            $table->integer('point');
             $table->decimal('monthly_residual');
             $table->timestamps();
         });
